Disabled crashing when the password is incorrect
The cause for the crashing was that the mail server could not be
accessed due to authenication failure. By replacing the "*****" password
with the actual password, this issue is resolved.

But I also added a try catch to raise a "Mail server unavailable"
message when a user tries to reset password while the application has
the wrong password.

This should solve StarcounterPrefabs/Registration#4 and the issues with
the restore password in People App.